tests: refactor and clean multipart upload test code
